# fea(ci): reduced slow test_diffusers timing. minor fixes