@imangohari1 2024-09-13
CC: @jiminha 
Here are the generated images of CI tests 
textual_inversion ([here](https://github.com/huggingface/optimum-habana/blob/main/tests/test_diffusers.py#L881))
 [
![pr1330_textual_inversion_pic](https://github.com/user-attachments/assets/6ddbc0eb-58fb-4ee1-8d5e-780af551a5d8)
](url)
train_controlnet ([here](https://github.com/huggingface/optimum-habana/blob/main/tests/test_diffusers.py#L2424)) 
![pr1330_train_controlnet_pic](https://github.com/user-attachments/assets/af02c9d7-ebbd-4f5c-bc0d-54db0e0aa476)