@libinta 2024-09-18
@imangohari1  we normally have the accuracy test at the end, with reduce of time, does the accuracy test pass?