@imangohari1 2024-09-18
> @imangohari1 we normally have the accuracy test at the end, with reduce of time, does the accuracy test pass?

@libinta 
Stable Diff tests/pipeline  do not  output accuracy at the end of the iteration. 
that is why I looked at the generated images here for accuracy. 
https://github.com/huggingface/optimum-habana/pull/1330#issuecomment-2350018532