@imangohari1 2024-09-25
@regisss 
ref: https://github.com/huggingface/optimum-habana/pull/1175#issuecomment-2330184808
I updated the latency value for `test_depth2img_pipeline_latency_bf16` since it has been failing since coded.